Added Gorodok
Gorodok is likely 99% percent fictional, but I've found an area in Russia which has similar patterns and matches the description of the map. Probably, devs used it as a foundation for the Gorodok map.
 
First, here are the link for two maps (Goggle and Yandex). Yandex is the most popular Russian map. Open either or both, but I suggest you use Yandex since that's where I found most of the hints:
https://yandex.eu/maps/-/CHGPNEn8
https://maps.app.goo.gl/8SG7R8xZJPnGq4pP7

To see the similarities I've found you first need to zoom out so you can see Kardymovo at the middle-left, Yartsevo at the top-right and Gorodok itself at the bottom-mid of your screen.

Now the findings:
1. The name of the settlement is Gorodok. Literally :)
2. It is located upon Dniepr river, exactly how it is said in the official map description. It's also in Smolensk region which also matches the description.
3. The path of the river there is somewhat similar to how the river in the game looks. It starts slightly curvy going from the bottom-left to the top right getting curlier and curlier. 
4. There is a settlement north to Gorodok called [Matryonino](https://yandex.eu/maps/-/CHGPVK5O) which could be the capture zone "Matrenino Hills" in the game. Geographically it's approximately in the same area as in the game.
5. There is a railway going from the bottom-left to the top-right. It also resembles the railway in the game.

Other than that, I haven't find anything. But I'm certain this is the area devs used as the foundation for the map and they made up a lot of things on top. Why?  Some places are completely out of this world.
Here are some examples:
1. People have been saying that the industrial site looks very unnatural, 3 buildings with 3 pipes - that's not how industrial sites usually look like.
2. The hilltop bunker. It is the remains of the bunker from WW2. If this place was real, then it would definitely be a local attraction, with a proper asphalt road to it. But in the game it looks like nobody except people living in the village nearby knows about it's existence.